pyproject.toml
[project]
name = "nne"
version = "0.1.0"
description = "customizable nne script"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "lxml>=5.3.2",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
]
